requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.103.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
# gmpy2==2.1.2
h11==0.14.0
idna==3.4
jax==0.4.26
jaxlib==0.4.26
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.4
mediapipe==0.10.9
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python>=4.7.0
opt-einsum==3.3.0
packaging==24.0
pillow==10.2.0
pip==23.3.1
protobuf==3.20.3
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
scipy==1.13.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
starlette==0.27.0
sympy==1.12
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
typing_extensions==4.11.0
urllib3==2.1.0
uvicorn==0.29.0
wheel==0.41.2
